+++ test/unit/test_print.gd
@@ -5,14 +5,19 @@
 extends "res://addons/gut/test.gd"
 
 var _old_indent_string = ''
+var _orig_err_as_fail = false
 
 func before_all():
 	gut.p('[before all]')
 	_old_indent_string = gut.logger.get_indent_string()
+	_orig_err_as_fail = gut.treat_error_as_failure
+
+	gut.treat_error_as_failure = false
 
 func after_all():
 	gut.p('[after all]')
 	gut.logger.set_indent_string(_old_indent_string)
+	gut.treat_error_as_failure = _orig_err_as_fail
 
 func before_each():
 	gut.p('[before each]')
@@ -53,7 +58,7 @@ func test_all_passing(p=use_parameters([[1, 2], [3, 4], [5, 6]])):
 
 func test_show_error():
 	_lgr.error('Something bad happened')
-	assert_false(false)
+	pass_test('look at error')
 
 func test_show_warning():
 	_lgr.warn('Something kinda bad happened')
@@ -142,15 +147,21 @@ class TestLogLevels:
 
 	var _orig_log_level = -1
 	var _orig_indent_string = null
+	var _orig_err_as_fail = false
 
 	func before_all():
 		_orig_log_level = gut.log_level
 		_orig_indent_string = gut.logger.get_indent_string()
+		_orig_err_as_fail = gut.treat_error_as_failure
+
+		gut.treat_error_as_failure = false
 		gut.logger.set_indent_string('--->')
+
 
 	func after_all():
 		gut.log_level = _orig_log_level
 		gut.logger.set_indent_string(_orig_indent_string)
+		gut.treat_error_as_failure = _orig_err_as_fail
 
 	func test_log_types_at_levels_with_passing_test(level=use_parameters([-2, -1, 0, 1, 2, 3])):
 		gut.log_level = level